backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
scikit-learn==1.3.2
numpy==1.26.2
Pillow==10.1.0
